+++ src/splat/__init__.py
@@ -1,7 +1,7 @@
 __package_name__ = __name__
 
 # Should be synced with pyproject.toml
-__version__ = "0.36.3"
+__version__ = "0.37.0"
 __author__ = "ethteck"
 
 from . import util as util